06_12_2020 Spring REST - Exception Handling Global Exception Handling with @ControllerAdvice. This is the best practice for large scale projects or realtime projects by making use of this @ControllerAdvice.
