+++ spring-rest-demo/src/main/java/de/ovsiannikov/springdemo/rest/StudentRestController.java
@@ -47,37 +47,4 @@ public class StudentRestController {
         }
         return theStudents.get(studentId);
     }
-
-    // Add an exception handler using @ExceptionHandler
-
-    @ExceptionHandler
-    public ResponseEntity<StudentErrorResponse> handlerException(StudentNotFoundException e) {
-
-        // create a StudentErrorResponse
-
-        StudentErrorResponse error = new StudentErrorResponse();
-
-        error.setStatus(HttpStatus.NOT_FOUND.value());
-        error.setMessage(e.getMessage());
-        error.setTimeStamp(System.currentTimeMillis());
-
-        // return ResponseEntity
-        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
-    }
-
-    // add another exception handler ... to catch any exception (catch all)
-
-    @ExceptionHandler
-    public ResponseEntity<StudentErrorResponse> handlerException(Exception e) {
-
-        // create a StudentErrorResponse
-        StudentErrorResponse error = new StudentErrorResponse();
-
-        error.setStatus(HttpStatus.BAD_REQUEST.value());
-        error.setMessage(e.getMessage());
-        error.setTimeStamp(System.currentTimeMillis());
-
-        // return ResponseEntity
-        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
-    }
 }
