+++ spring-rest-demo/src/main/java/de/ovsiannikov/springdemo/rest/StudentRestExceptionHandler.java
@@ -1,0 +1,45 @@
+package de.ovsiannikov.springdemo.rest;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class StudentRestExceptionHandler {
+
+    // add exception handling code here
+
+    // Add an exception handler using @ExceptionHandler
+
+    @ExceptionHandler
+    public ResponseEntity<StudentErrorResponse> handlerException(StudentNotFoundException e) {
+
+        // create a StudentErrorResponse
+
+        StudentErrorResponse error = new StudentErrorResponse();
+
+        error.setStatus(HttpStatus.NOT_FOUND.value());
+        error.setMessage(e.getMessage());
+        error.setTimeStamp(System.currentTimeMillis());
+
+        // return ResponseEntity
+        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+    }
+
+    // add another exception handler ... to catch any exception (catch all)
+
+    @ExceptionHandler
+    public ResponseEntity<StudentErrorResponse> handlerException(Exception e) {
+
+        // create a StudentErrorResponse
+        StudentErrorResponse error = new StudentErrorResponse();
+
+        error.setStatus(HttpStatus.BAD_REQUEST.value());
+        error.setMessage(e.getMessage());
+        error.setTimeStamp(System.currentTimeMillis());
+
+        // return ResponseEntity
+        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+    }
+}
